// Source/DashboardTabApp/src/app/models/Course.ts
export class Course {
    id: number = 0;
    name: string = "";
    predictiveQnAServiceHost: string = "";
    predictiveQnAKnowledgeBaseId: string = "";
    predictiveQnAEndpointKey: string = "";
    predictiveQnAHttpEndpoint: string = "";
    predictiveQnAHttpKey: string = "";
    predictiveQnAKnowledgeBaseName: string = "";
    predictiveQnAConfidenceThreshold: string = "";
    deployedURL: string = "";
    groupId: string = "";
    edited: boolean = false; // true if edited or new
}